# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Open-Kanban API
  description: |
    Open-Kanban is a self-hosted kanban board API server.
    
    ## Authentication
    
    This API uses Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-token>
    ```
    
    Some endpoints (marked as public) do not require authentication.
    
    ## Signature Verification (IMPORTANT SECURITY INFORMATION)
    
    Protected endpoints support HMAC-SHA256 signature verification for enhanced security. However, **signature verification is DISABLED by default**.
    
    ### Security Warning
    
    When signature verification is disabled (the default), protected endpoints will accept requests without signature verification, even if they are marked as requiring signature verification in this documentation.
    
    ### To Enable Signature Verification
    
    Set the following environment variables:
    
    1. `SIGNATURE_ENABLED=1` - Enable signature verification
    2. `SIGNATURE_SECRETS=access-key-1:secret-key-1,access-key-2:secret-key-2` - Configure access keys and their corresponding secrets (comma-separated key:value pairs)
    
    ### How Signature Verification Works
    
    When enabled, protected endpoints require the following headers:
    
    - `X-Signature`: HMAC-SHA256 signature (hex-encoded)
    - `X-Timestamp`: Unix timestamp of the request (must be within 5 minutes of server time)
...
    
    ### Signature Generation
    
...
    
...
servers:
...
paths:
...
  /tasks:
    get:
      summary: Get tasks
      description: Public endpoint to get tasks with optional filters
      operationId: getTasks
      tags:
        - Tasks
      parameters:
        - name: columnId
          in: query
          schema:
            type: string
        - name: boardId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: List of tasks with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  total:
                    type: integer
                  page:
                    type: integer
                  pageSize:
                    type: integer
                  pageCount:
                    type: integer
    post:
      summary: Create a task
      description: Create a new task in a column
      operationId: createTask
      tags:
        - Tasks
      security:
        - BearerAuth: []
        - SignatureAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '200':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          description: Rate limit exceeded
  /tasks/{id}:
    get:
      summary: Get a task by ID
      description: Public endpoint to get task details
      operationId: getTask
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: include
          in: query
          description: Include additional data (comments, subtasks)
          schema:
            type: string
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a task
      description: Update task properties
      operationId: updateTask
      tags:
        - Tasks
      security:
        - BearerAuth: []
        - SignatureAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a task
      description: Delete a task
      operationId: deleteTask
      tags:
        - Tasks
      security:
        - BearerAuth: []
        - SignatureAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    SignatureAuth:
      type: apiKey
      in: header
      name: X-Signature
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        assignee:
          type: string
        meta:
          type: object
        columnId:
          type: string
        position:
          type: integer
        published:
          type: boolean
        archived:
          type: boolean
        archivedAt:
          type: string
          format: date-time
        agentId:
          type: string
        agentPrompt:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        commentCount:
          type: integer
        subtaskCount:
          type: integer
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        subtasks:
          type: array
          items:
            $ref: '#/components/schemas/Subtask'
    CreateTaskRequest:
      type: object
      required:
        - title
        - columnId
      properties:
        title:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 5000
        priority:
          type: string
          enum: [low, medium, high]
        assignee:
          type: string
          maxLength: 100
        meta:
          type: object
        columnId:
          type: string
        position:
          type: integer
        published:
          type: boolean
        agentId:
          type: string
        agentPrompt:
          type: string
          maxLength: 2000
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 5000
        priority:
          type: string
          enum: [low, medium, high]
        assignee:
          type: string
          maxLength: 100
        meta:
          type: object
        columnId:
          type: string
        position:
          type: integer
        published:
          type: boolean
        agentId:
          type: string
        agentPrompt:
          type: string
          maxLength: 2000
    Comment:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        author:
          type: string
        taskId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Subtask:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        completed:
          type: boolean
        taskId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Not logged in
    Forbidden:
      description: Forbidden - User does not have permission
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: No permission
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Resource not found
...